Dirty first attempt to plug the libzim inside kiwix-js.
For some reason, the libzim can only be called once. So it's only used
for the first request.
And you have to wait for the "libzim initialized" in browser console
before displaying an article.
It only works on ZIM files with size <2GB.
On some ZIM files, the HTML string returned by libzim seems to be a
concatenation of many articles instead of only the expected article, as if it
was continuing to read the ZIM file content after the article end.